Implementing systemcall for shotdown
This commit implements a systemcall for shutdown the system.
It provides the necessary systemcall which creates a new thread
for the kernel task. The thread will then call halt(), reboot()
or cancel the pending shutdown process. The systemcall takes
three parameters: a modus (cancel, halt, reboot), a delay and
a boolean for changing into the kernel console