Do not block faulted task until debugger exits
If a task faults while a debugger (such as trace) is attached, we must
not block the task until the debugger exits (trace waits for task to
exit). It is sufficient to wait for a debugger to start, because
the task gets blocked in udebug_stoppable_end() (since the task is
not GO).