Volsrv should not use zero as a valid reference count
Volume server allows the reference count for a valid volume with
persistent configuration to be zero, but refcnt did not expect that
and increasing refcount up from zero tripped assertion in the same_twice
test. The solution is to add one extra reference to persistent volumes
and have zero reference count always mean the volume can be freed.