Correcting `sysctl list-units`
One failure was a cast in lib/sysman/src/ctl.c which wrongly used
an enum like a pointer of the type sysarg_t. This resulted into a
memory allocation failure.

The second bug was in the loop of sysctl. For iterating over all
found units the difference was created of two pointers and then
compared to an unsigned int which was down casted to an int.
Unnecessary complicated and a source for casting problems.
It has been replaced with a simple for-loop and an array access