Remove XXX comment from each call to display_get_gc()
These were added with the intent that each piece of code that does rendering
in the display server is augmented with a loop over all display devices.
The better solution seems to be to introduce a virtualization layer in the
form of a cloning GC that clones the rendering operations to each display
device. It makes the business code simpler and separates concerns better.
It could also allow having more types of output GCs, than just display
devices (such as an 'observer' for RFB).