Repaint desktop when window is brought to front
Libgui used to repaint the entire window when it received focus so this
went unnoticed. Libui currently only repaints the window decoration.
Repainting the entire desktop is probably not necessary, but I am
mimicking the other places where we do this (e.g. moving window).
All of these cases can be revisited and, perhaps, optimized to
repaint just a part of the desktop.