Remember current position in libriff and optimize sequential reading
TPF files currently contain a lot of chunks and each chunk means
a ftell/fseek. libc's ftell/fseek flush the I/O buffer, leading
to a lot of unnecessary IPC. This all leads to reading typeface to
take way too long.

We remember the current file position
in a RIFF reader and avoid ftell and also avoid fseek at the end
of a chunk, if possible (if we need to advance just a little bit,
we read instead of seek).