Switch to previous (last) window when cycling
Focusing a window brings it to top, i.e. first in the list (the order of
which is the stacking order). Thus the formerly first window becomes
the second. Switching to the next window thus always cycles trough
the same two windows. Switching to the previous (actualy last window,
there is no previous window) will cycle through all the windows.