Fix build with meson 0.56.2
Meson changed the meaning of warning_level, before 2 had -Wpedantic
and 3 was -Wextra, now -3 is -Wpedantic and 2 is -Wextra. HelenOS
can compile with -Wextra, but not with -Wpedantic (and lot of work
to make it compile with that). So changing to level 3 for now.
This probably means build will fail with older meson version.