Go To Line dialog
Need to restore cursor after closing dialog. This is not how it should
be done. Either the cursor should be restored by pane activation
handler (so we still treat cursor as a global resource and the
individual controls need to responsibly manage it), or, the UI library
should somehow virtualize it, e.g. each window would remember its
cursor state and cursor would be automatically updated when a different
window is activated.