Make sure page is not beyond end of list
Need to correct max_idx correctly for short lists (was getting a
negative number and putting it to a size_t). Need to prevent
page going beyond end of list when moving up a dir and seeking
to the entry of directory we just exited (unless the list is short).