Add -c option to terminal to execute a command
This is to allow terminal-mode commands to be run from non-terminal
(i.e. GUI) context. This way is similar to what UN*X style OSses do.
For the record, I think this is a kludge. I think we can do better.
A console/terminal window could be automatically open when needed
(something like that happens in Windows?)