Fix reading and writing chunk padding
You cannot use riff_read() to read the chunk padding, because it's not
part of the chunk data proper. This fixes libgfxfont's unit test, which
could not read back the TPF it wrote.

Also, it seems prudent to actually write the padding, not just seek
over it. (Might not be a problem but, just in case...)