Pause before resuming UI, if needed
To be able to see output of commands executed from Navigator, we need
to pause before repainting the UI. This should not be needed, however,
for binaries that do not leave anything on the screen (e.g. (G)UI
applications). We use a simple heuristic - if the cursor is in the top
left corner of the screen, we don't pause.