add relative bottom left/right taskbar coordinates
the absolute coordinates of the screen and relative to the taskbar have been added both in the lower left and in the lower right.
Absolute coordinates place the window above the taskbar, while non-absolute coordinates place the window above the taskbar.
The taskbar delta calculation is a fixed value that corresponds to the height of the taskbar. It is necessary to create a communication the size of the taskbar.