Eliminate an undefined behavior in scheduler debug printout
The .relink_deadline field is only normally accessed by the local CPU
and is not synchronized in any way. This is technically undefined
behavior, even if the only realistic failure mode I can think of here
is the output value being nonsense. Because the printed out value has
next to no utility to anyone using the command, simply remove it
instead of ignoring the issue or making the variable atomic.