fix calculation of TLS size for main fibril
Before this patch, _tcb_data_offset always used __progsymbols.elfstart. However,
that is wrong when it is being called from the loader server!! Now we pass to it
a pointer to the correct ELF, falling back to elfstart in the public tls_get call.

(debugging this was quite an ordeal and took me like 5 hours, thanks for asking (':)