Use tar to install toolchains into the destination
Using a pair of tar invocations (reading the source path by one
instance and piping the data to a second instance that writes into
the target path) has the benefit of better preserving certain file
attributes and especially the hard links (even if the source and the
target paths reside on different physical file systems).

Preserving the hard links is important to avoid needlessly wasting
storage space.